==> Data Analysis/P1 Filtering.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

%cd /content/drive/MyDrive/Processed

!pip install pandas lxml
import pandas as pd
import xml.etree.ElementTree as ET
import glob
import os

input_file = '/content/drive/MyDrive/Processed/Processed1/final_merged_cleaned_questions.csv'
output_folder = '/content/drive/MyDrive/Processed/Processed1/'

os.makedirs(output_folder, exist_ok=True)

def filter_by_tags(df, react_tags):
    return df[df['Tags'].notnull() & df['Tags'].str.contains('|'.join(react_tags), case=False)]

react_tags = [
    'reactjs', 'react-native', 'react-redux', 'react-router', 'react-hooks',
    'react-navigation', 'react-bootstrap', 'react-native-ios', 'react-props',
    'react-native-flatlist', 'react-apollo', 'react-testing-library',
    'react-select', 'react-dom', 'reactjs-flux', 'create-react-app',
    'reactjs.net', 'reactjs-testutils', 'konvajs-reactjs', 'reactjs-native',
    'video-reactjs', 'reactjs-popup', 'applicationinsights-react-js'
]

df = pd.read_csv(input_file)

filtered_df = filter_by_tags(df, react_tags)

output_filtered_file = os.path.join(output_folder, 'final_filtered_questions.csv')

filtered_df.to_csv(output_filtered_file, index=False)
print(f'Successfully filtered and saved to: {output_filtered_file}')

questions_df = filtered_df[filtered_df['PostTypeId'] == 1]
answers_df = filtered_df[filtered_df['PostTypeId'] == 2]

print(f"Jumlah questions: {len(questions_df)}")
print(f"Jumlah answers: {len(answers_df)}")

if not questions_df.empty:
    questions_df = questions_df.drop_duplicates()
    cleaned_questions_file = os.path.join(output_folder, 'cleaned_questions_final.csv')
    questions_df.to_csv(cleaned_questions_file, index=False)
    print(f"  -> Cleaned questions saved to: {cleaned_questions_file}")

if not answers_df.empty:
    answers_df = answers_df.drop_duplicates()
    cleaned_answers_file = os.path.join(output_folder, 'cleaned_answers_final.csv')
    answers_df.to_csv(cleaned_answers_file, index=False)
    print(f"  -> Cleaned answers saved to: {cleaned_answers_file}")
