==> Data Analysis/P2 addColumns.ipynb <==
!pip install pandas lxml
import pandas as pd
import xml.etree.ElementTree as ET
import os
import re
import glob

pip install --upgrade pip

!pip install pandas lxml
import pandas as pd
import xml.etree.ElementTree as ET
import os
import re
import glob

from google.colab import drive
drive.mount('/content/drive')

%cd /content/drive/MyDrive/Processed

input_file = '/content/drive/MyDrive/Processed/Processed2/cleaned_questions_final.csv'
output_folder = '/content/drive/MyDrive/Processed/Processed2/'

df = pd.read_csv(input_file)

df.head()

df['code_snippet'] = df['Body'].apply(lambda x: 1 if re.search(r'<code>.*?</code>', str(x), re.DOTALL) else 0)

print(df[['Body', 'code_snippet']].head())

df['question_line_count'] = df['Body'].apply(lambda x: len(str(x).splitlines()))

def count_code_lines(text):
    code_blocks = re.findall(r'<code>(.*?)</code>', str(text), re.DOTALL)
    return sum(len(code_block.splitlines()) for code_block in code_blocks)

df['code_line_count'] = df['Body'].apply(count_code_lines)

print(df[['Body', 'question_line_count', 'code_line_count']].head())

df['image'] = df['Body'].apply(lambda x: 1 if re.search(r'<img\s+[^>]*src=', str(x)) else 0)

print(df[['Body', 'image']].head())

df['AnswerCount'] = pd.to_numeric(df['AnswerCount'], errors='coerce').fillna(0)
df['ViewCount'] = pd.to_numeric(df['ViewCount'], errors='coerce').fillna(0)

df['pd_score'] = (df['AnswerCount'] / df['ViewCount']) * 100

print(df[['Body', 'pd_score']].head())

output_file = '/content/drive/MyDrive/Processed/Processed2/addcolumns_final_merged_cleaned_questions.csv'
df.to_csv(output_file, index=False, sep=',', quotechar='"', encoding='utf-8')

print(f"File berhasil disimpan ke {output_file}")

file_path = '/content/drive/MyDrive/Processed/Processed2/addcolumns_final_merged_cleaned_questions.csv'
df = pd.read_csv(file_path)

df.head()
