==> Data Analysis/P3 userReputation.ipynb <==
import os
import pandas as pd

users_split_path = 'E:/source/users_split/'
questions_file_path = 'E:/progres2/addcolumns_final_merged_cleaned_questions.csv'

questions_df = pd.read_csv(questions_file_path)

user_files = [os.path.join(users_split_path, f) for f in os.listdir(users_split_path) if f.endswith('.csv')]

list_of_user_dfs = [pd.read_csv(file) for file in user_files]

users_df = pd.concat(list_of_user_dfs, ignore_index=True)

print(questions_df.head())

print(users_df.head())

users_df.rename(columns={'Id': 'OwnerUserId'}, inplace=True)

merged_df = pd.merge(questions_df, users_df[['OwnerUserId', 'Reputation']], on='OwnerUserId', how='left')

merged_df.head()

def categorize_reputation(reputation):
    if reputation >= 2400:
        return 'high'
    elif 400 <= reputation < 2400:
        return 'mid'
    elif 1 < reputation < 400:
        return 'low'
    else:
        return 'unknown'

merged_df['ReputationCategory'] = merged_df['Reputation'].apply(categorize_reputation)

print(merged_df[['OwnerUserId', 'Reputation', 'ReputationCategory']].head())

merged_df.head()

output_file_path = 'E:/output_addcolumns_final_merged_cleaned_questions.csv'
merged_df.to_csv(output_file_path, index=False)

import pandas as pd
df = pd.read_csv("E:/output_addcolumns_final_merged_cleaned_questions.csv")
df.head()

df[df["code_snippet"] == 1].head()

df[df["AnswerCount"] == 0].head()

df.columns

df[df["PostTypeId"] == 2].head()
