==> Data Analysis/P4 addAnswered.ipynb <==
from google.colab import drive
drive.mount('/content/drive')

%cd /content/drive/MyDrive/Processed/Processed3

import pandas as pd
df = pd.read_csv("output_addcolumns_final_merged_cleaned_questions.csv")

df.head()

df['answered?'] = df['AnswerCount'].apply(lambda x: 1 if x > 0 else 0)

print("Kolom 'answered?' berhasil ditambahkan.")

df.head()

df[df["answered?"] == 0]

columns_to_drop = ['PostTypeId', 'AcceptedAnswerId', 'OwnerUserId', 'AnswerCount',
                   'FavoriteCount', 'CommunityOwnedDate', 'CreationDate', 'Score', 'Title', 'Body']
df.drop(columns=columns_to_drop, inplace=True)

df['Reputation'] = pd.to_numeric(df['Reputation'], errors='coerce')

df = df[df['Reputation'] > 0].dropna(subset=['Reputation'])

def categorize_reputation(reputation):
    if reputation >= 2400:
        return 'High'
    elif 400 <= reputation < 2400:
        return 'Mid'
    elif 1 < reputation < 400:
        return 'Low'
    else:
        return None

df['ReputationCategory'] = df['Reputation'].apply(categorize_reputation)

df = df.dropna(subset=['ReputationCategory'])

df.head()

missing_or_zero = df['Reputation'].isna().sum() + (df['Reputation'] == 0).sum()
print(f"Jumlah nilai 0 atau missing values di kolom 'Reputation': {missing_or_zero}")

output_path = '/content/drive/My Drive/Processed/Processed3/cleaned_output_addcolumns_final_merged_cleaned_questions.csv'
df.to_csv(output_path, index=False)

print("Dataset berhasil dibersihkan dan disimpan.")

import pandas as pd
df2 = pd.read_csv("cleaned_output_addcolumns_final_merged_cleaned_questions.csv")
print(df2.head())

df2.head()
