==> Data Analysis/P8 randomSampling.ipynb <==
import pandas as pd
import math

from google.colab import drive
drive.mount('/content/drive')
%cd /content/drive/MyDrive/Processed/Processed4
df = pd.read_csv("pengaruhReputasi.csv")
df.head()

confidence_level = 0.99
margin_error = 0.01
population_size = 534820
p = 0.5
Z = 2.576  # Untuk confidence level 99%

n0 = (Z**2 * p * (1 - p)) / (margin_error**2)
print(f"Initial sample size (n0): {n0}")

n_final = n0 / (1 + ((n0 - 1) / population_size))
n_final = math.ceil(n_final)
print(f"Final sample size after finite population correction: {n_final}")

actual_population = len(df)
print(f"Jumlah total data di CSV: {actual_population}")

if n_final > actual_population:
    print("Sample size exceeds population. Use full population size.")
    n_final = actual_population

sample_df = df.sample(n=n_final, random_state=42)
print(f"Number of data in the sample: {len(sample_df)}")
print(sample_df.head())

sample_df.to_csv('pengaruhReputasi_sampled.csv', index=False)
print("The sample has been saved to 'pengaruhReputasi_sampled.csv'")
